Refractor axios instance
There was an issue with importing an instance from a separate file as it resulted in a big lag delay when you make subsequent enqueue/dequeue post requests. The issue was isolated to the axios instance from a separate file therefore the conents were just moved for this paricular request functions.