Refractor student to user state
Previously was student state.  Now almost all actions are universal (for recruiters and students). There are a couple of differentiations such as the success action which changes the state differently. The student is not expected to get the credentials from the user while the company is. Also isStudent property added.